feat: Add bounding box to map API URL
This commit adds support for displaying a bounding box on the map in the `App` component. It introduces a new `boundingBoxPath` variable that generates the path for the bounding box based on the `location.boundingbox` array. The bounding box is drawn as a red polygon on the map, highlighting the area of the location. This change enhances the visual representation of the location and provides additional context to the user.