feat: Add movies API integration to App component
This commit adds the integration of the movies API to the `App` component. It introduces a new state variable `movies` and a new function `getMovies` that fetches movie data based on the location. The movies data is then displayed in the UI using the `Movies` component. This enhancement expands the functionality of the City Explorer application by providing information about movies related to the searched location.